refactor: Move common layman types under types folder
