+++ projects/hslayers/src/common/layman/public-api.ts
@@ -3,7 +3,8 @@ export * from './layman-current-user.component';
 export * from './layman-login.component';
 export * from './layman.module';
 export * from './layman.service';
-export * from './current-user-response.type';
-export * from './post-layer-response.type';
-export * from './get-layers-response.type';
-export * from './delete-layer-response.type';
+export * from './types/current-user-response.type';
+export * from './types/post-layer-response.type';
+export * from './types/get-layers-response.type';
+export * from './types/delete-layer-response.type';
+export * from './types/layman-user.type';
