+++ projects/hslayers/src/common/layman/types/current-user-response.type.ts
