+++ projects/hslayers/src/common/layman/types/delete-layer-response.type.ts
@@ -1,4 +1,4 @@
-import {accessRightsModel} from './../../components/add-data/common/access-rights.model';
+import {accessRightsModel} from './../../../components/add-data/common/access-rights.model';
 
 export type DeleteAllLayersResponse = {
   access_rights?: accessRightsModel;
